add servicehandler interface
